+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fixed pipeline filter [#414](https://github.com/matchms/matchms/pull/414)
+- Removed fingerprint writing to file [#416](https://github.com/matchms/matchms/pull/416)
 
 ### Fixed